Split the setup process into two possiblilities
This allows the user to determine if they want to manually select the
LTSpice installation on their system or if they want the ULP to try to
set it up automatically. Still have to figure out a nice way to setup
the rootpath variable in a nice OS independent way.