Created assign_rootpath function
The goal of this function is to abstract the rootpaths from the main ulp
and move them to the include ulp. The code is not fully tested and
working under Linux. The only issue I have is that it can't find ANY of
the executables under Linux automatically. Still have to look into that.